requirements.txt
coqui-tts>=0.22.0
torch>=2.0.0
openai-whisper>=20231117
anthropic>=0.18.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
